requirements.txt
llama-cpp-python==0.1.36; platform_system != "Windows"